Unbreak MaltParser.
Maltparser will only work when the model file is in the current working
directory. The old maltparser wrapper script did that (although the reason why
wasn't terribly informative), and the recent commit that removed that script
broke this.

This commit updates the Python code to chdir() to the correct dir before
invoking the parser.